requirements.txt
streamlit==1.22.0
tensorflow==2.11.0
joblib==1.3.1
Pillow==9.1.1
numpy==1.23.0
requests==2.31.0
